refactor: modularize main.py into FastAPI APIRouters
