+++ backend/dependencies.py
@@ -1,0 +1,76 @@
+import os
+import sys
+from pathlib import Path
+from dotenv import load_dotenv
+from slowapi import Limiter
+from slowapi.util import get_remote_address
+
+limiter = Limiter(key_func=get_remote_address)
+
+# Load environment variables
+env_path = Path(__file__).parent / '.env'
+load_dotenv(dotenv_path=env_path)
+
+# Initialize Supabase Client
+try:
+    from supabase import create_client, Client
+    url = os.environ.get("SUPABASE_URL")
+    key = os.environ.get("SUPABASE_SERVICE_KEY")
+    if not url or not key:
+        print("[ERROR] SUPABASE_URL or SUPABASE_SERVICE_KEY not set in backend/.env")
+        supabase = None
+    else:
+        supabase = create_client(url, key)
+except (ImportError, Exception) as e:
+    print(f"[WARNING] Supabase initialization failed: {e}")
+    supabase = None
+    Client = None
+
+# Ensure project root is on path for imports
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
+
+from backend.services.classifier_service import ClassifierService
+from backend.services.classifier_v2 import classifier_v2
+from backend.services.classifier_v3 import classifier_v3
+from backend.services.ner_service import NERService
+from backend.services.duplicate_service import DuplicateService
+from backend.services.rag_service import RagService
+
+# Initialize singletons
+classifier_service = ClassifierService()
+ner_service = NERService()
+duplicate_service = DuplicateService()
+rag_service = RagService()
+
+try:
+    from backend.services.gemini_service import GeminiService
+    gemini_service = GeminiService()
+except ImportError:
+    gemini_service = None
+
+try:
+    from backend.services.ocr_service import OCRService
+    ocr_service = OCRService()
+except ImportError:
+    ocr_service = None
+
+# Shared In-Memory DB
+TICKETS_DB = []
+
+def get_system_settings(company_id: str) -> dict:
+    defaults = {
+        "ai_confidence_threshold": 0.80,
+        "duplicate_sensitivity": 0.85,
+        "enable_auto_resolve": False
+    }
+    if not supabase or not company_id:
+        return defaults
+    try:
+        res = supabase.table("system_settings").select(
+            "ai_confidence_threshold, duplicate_sensitivity, enable_auto_resolve"
+        ).eq("company_id", company_id).single().execute()
+        if res.data:
+            return {**defaults, **res.data}
+    except Exception as e:
+        print(f"[WARNING] Could not fetch system_settings for company_id={company_id}: {e}")
+    return defaults
